Add helpers file and update tests
